选择类型

+++ litemall-db/sql/social_table.sql
@@ -648,9 +648,9 @@ CREATE TABLE `litemall_ad_rent` (
 
 DROP TABLE IF EXISTS `litemall_type_obj`;
 CREATE TABLE `litemall_type_obj` (
-  `id` tinyint(3) NOT NULL AUTO_INCREMENT,
+  `id` int(11) NOT NULL AUTO_INCREMENT,
   `name` varchar(100) NOT NULL COMMENT '回答内容',
-  `is_show_client` tinyint(3) NOT NULL DEFAULT '0' COMMENT '客户端是否显示',
+  `is_show_client` tinyint(1) NOT NULL DEFAULT '0' COMMENT '客户端是否显示',
   `add_time` datetime DEFAULT NULL COMMENT '创建时间',
   `update_time` datetime DEFAULT NULL COMMENT '创建时间',
   PRIMARY KEY (`id`)
